reorganized files that came from app startup
