+++ FrugalApp/components/react_native_basics/Collapsible.tsx
@@ -1,9 +1,9 @@
 import { PropsWithChildren, useState } from 'react';
 import { StyleSheet, TouchableOpacity } from 'react-native';
 
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { IconSymbol } from '@/components/ui/IconSymbol';
+import { ThemedText } from '@/components/react_native_basics/ThemedText';
+import { ThemedView } from '@/components/react_native_basics/ThemedView';
+import { IconSymbol } from '@/components/react_native_basics/ui/IconSymbol';
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
