+++ FrugalApp/components/react_native_basics/HapticTab.tsx
