+++ FrugalApp/components/react_native_basics/HelloWave.tsx
@@ -8,7 +8,7 @@ import Animated, {
   withSequence,
 } from 'react-native-reanimated';
 
-import { ThemedText } from '@/components/ThemedText';
+import { ThemedText } from '@/components/react_native_basics/ThemedText';
 
 export function HelloWave() {
   const rotationAnimation = useSharedValue(0);
