+++ FrugalApp/components/react_native_basics/ThemedText.tsx
