+++ FrugalApp/components/react_native_basics/ThemedView.tsx
