+++ FrugalApp/components/react_native_basics/ui/IconSymbol.ios.tsx
