+++ FrugalApp/components/react_native_basics/ui/TabBarBackground.ios.tsx
