+++ FrugalApp/components/react_native_basics/ui/TabBarBackground.tsx
